python/requirements.txt
aiofiles==23.2.1
aiogram==2.21
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.7.0
async-timeout==4.0.3
attrs==24.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
bidict==0.22.1
billiard==4.2.0
blinker==1.8.2
cachetools==5.3.1
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.0
customtkinter==5.2.2
darkdetect==0.8.0
distro==1.9.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Mail==0.10.0
flask-marshmallow==1.2.1
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.5
frozenlist==1.4.1
gensim==4.3.3
google-api-core==2.12.0
google-api-python-client==2.22.0
google-auth==2.22.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
googletrans==3.0.0
greenlet==3.0.0
gspread==5.10.0
gunicorn==20.1.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jwt==1.3.1
keyboard==0.13.5
kombu==5.3.7
magic-filter==1.0.12
MarkupSafe==2.1.5
marshmallow==3.21.3
multidict==6.0.5
nltk==3.9.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
olefile==0.47
openai==1.57.4
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
prompt_toolkit==3.0.47
protobuf==4.24.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.1
pystray==0.19.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.7.0
python-socketio==5.8.0
pytz==2024.1
pywin32==306
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
schedule==1.2.0
scipy==1.13.1
six==1.16.0
slack_bolt==1.19.1
slack_sdk==3.31.0
smart-open==7.0.5
sniffio==1.3.1
soupsieve==2.4.1
SQLAlchemy==2.0.31
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==3.0.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.17.0
yarl==1.9.4
